untitled folder/requirements.txt
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.13
ci-info==0.3.0
click==8.2.1
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==3.1.2
Django==5.2.3
django-cors-headers==4.9.0
djangorestframework==3.16.0
durationpy==0.10
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.26
langchain-openai==0.3.27
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.2
mypy_extensions==1.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.0
open_clip_torch==2.32.0
openai==1.93.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
posthog==6.0.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.16
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
